content -u

+++ Back_End/src/main/java/lk/ijse/spring/entity/Payment.java
@@ -21,7 +21,7 @@ public class Payment {
     private String paymentId;
     private LocalDate date;
     private double rent_amount;
-    private long extra_mileage;
+    private double extra_mileage;
     private double cost_per_extra_km;
     private double damage_cost;
     private String damageDescription;
